+++ R/tools.R
@@ -79,6 +79,10 @@ check_paths <- function(paths) {
 #' with the `legacy_max_lint_r_errors` option in `rhino.yml`.
 #' This can be useful when inheriting legacy code with multiple styling issues.
 #'
+#' The [box.linters::namespaced_function_calls()] linter requires the `{treesitter}` and
+#' `{treesitter.r}` packages. These require R >= 4.3.0. `lint_r()` will continue to run and skip
+#' `namespaced_function_calls()` if its dependencies are not available.
+#'
 #' @param paths Character vector of directories and files to lint.
 #' When `NULL` (the default), check `app` and `tests/testthat` directories.
 #'
@@ -87,6 +91,17 @@ check_paths <- function(paths) {
 #' @export
 # nolint end
 lint_r <- function(paths = NULL) {
+  if (!box.linters::is_treesitter_installed()) {
+    cli::cli_warn(
+      c(
+        "!" = paste(
+          "`box.linters::namespaced_function_calls()` requires {{treesitter}} and {{treesitter.r}}",
+          "to be installed."
+        ),
+        "i" = "`lintr_r()` will continue to run using the other linter functions."
+      )
+    )
+  }
   if (is.null(paths)) {
     paths <- c("app", "tests/testthat")
   }
@@ -122,10 +137,13 @@ rhino_style <- function() {
 
 #' Format R
 #'
-#' Uses the `{styler}` package to automatically format R sources.
+#' Uses the `{styler}` and `{box.linters}` packages to automatically format R sources.
 #'
 #' The code is formatted according to the `styler::tidyverse_style` guide with one adjustment:
 #' spacing around math operators is not modified to avoid conflicts with `box::use()` statements.
+#'
+#' `box.linters::style_*` functions require the `treesitter` and `treesitter.r` packages. These, in
+#' turn, require R >= 4.3.0.
 #'
 #' @param paths Character vector of files and directories to format.
 #' @return None. This function is called for side effects.
@@ -140,10 +158,21 @@ rhino_style <- function() {
 #' }
 #' @export
 format_r <- function(paths) {
+  if (!box.linters::is_treesitter_installed()) {
+    cli::cli_abort(
+      c(
+        "x" = "The packages {{treesitter}} and {{treesitter.r}} are required by `format_r()`",
+        "i" = "These package require R version >= 4.3.0 to install."
+      )
+    )
+  }
+
   for (path in paths) {
     if (fs::is_dir(path)) {
+      box.linters::style_box_use_dir(path)
       styler::style_dir(path, style = rhino_style)
     } else {
+      box.linters::style_box_use_file(path)
       styler::style_file(path, style = rhino_style)
     }
   }